iox-#1345 Replace old logger with new one
