+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/log/platform_building_blocks/console_logger.inl
@@ -18,7 +18,7 @@
 #ifndef IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_CONSOLE_LOGGER_INL
 #define IOX_HOOFS_PLATFORM_BUILDING_BLOCKS_CONSOLE_LOGGER_INL
 
-#include "iceoryx_hoofs/log_ng/platform_building_blocks/console_logger.hpp"
+#include "iceoryx_hoofs/log/platform_building_blocks/console_logger.hpp"
 
 namespace iox
 {
