+++ iceoryx_hoofs/include/iceoryx_hoofs/log/platform_building_blocks/logcommon.hpp
