--- iceoryx_hoofs/source/log/hoofs_logging.cpp
@@ -1,24 +1,0 @@
-// Copyright (c) 2022 by Apex.AI Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#include "iceoryx_hoofs/internal/log/hoofs_logging.hpp"
-
-namespace iox
-{
-constexpr char LoggingComponentHoofs::Ctx[];
-constexpr char LoggingComponentHoofs::Description[];
-
-} // namespace iox
