--- iceoryx_hoofs/source/log/logging_internal.cpp
@@ -1,29 +1,0 @@
-// Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#include "logging_internal.hpp"
-
-namespace iox
-{
-namespace log
-{
-// NOLINTNEXTLINE(readability-identifier-naming)
-constexpr char LoggingComponent::Ctx[];
-// NOLINTNEXTLINE(readability-identifier-naming)
-constexpr char LoggingComponent::Description[];
-} // namespace log
-} // namespace iox
