+++ iceoryx_hoofs/source/log/platform_building_blocks/console_logger.cpp
@@ -15,7 +15,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/log_ng/platform_building_blocks/console_logger.hpp"
+#include "iceoryx_hoofs/log/platform_building_blocks/console_logger.hpp"
 #include "iceoryx_hoofs/platform/time.hpp"
 
 #include <cstdio>
