+++ iceoryx_hoofs/source/log/platform_building_blocks/logger.cpp
@@ -16,8 +16,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "iceoryx_hoofs/log_ng/platform_building_blocks/logger.hpp"
-#include "iceoryx_hoofs/log_ng/platform_building_blocks/logcommon.hpp"
+#include "iceoryx_hoofs/log/platform_building_blocks/logger.hpp"
+#include "iceoryx_hoofs/log/platform_building_blocks/logcommon.hpp"
 
 #include <atomic>
 #include <iomanip>
